backend/requirements.txt
annotated-types==0.7.0
base58==2.1.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cytoolz==1.1.0
ecdsa==0.19.1
eth-hash==0.7.1
eth-keys==0.7.0
eth-typing==5.2.1
eth-utils==5.3.1
idna==3.11
more-itertools==10.8.0
py-bip39-bindings==0.3.0
py-ed25519-zebra-bindings==1.3.0
py-sr25519-bindings==0.2.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyNaCl==1.6.0
python-dotenv==1.2.1
requests==2.32.5
scalecodec==1.2.12
six==1.17.0
substrate-interface==1.7.11
tapo==0.8.7
toolz==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
xxhash==3.6.0
